fix bugs and tests
