+++ src/cli/commands/index.ts
@@ -4,10 +4,8 @@
 
 export * from "./deposit";
 export * from "./eth1-private-network";
-
 export * from "./beacon";
 export * from "./create-config";
 export * from "./wallet";
-export * from "./start";
 export * from "./validator";
