+++ test/unit/validator/validator.test.ts
@@ -1,11 +1,11 @@
 import {expect} from "chai";
 import Validator from "../../../src/validator";
-import {ValidatorCtx} from "../../../src/validator/types";
-import {RpcClient, RpcClientOverInstance} from "../../../src/validator/rpc";
+import {RpcClientOverInstance} from "../../../src/validator/rpc";
 import {Keypair} from "@chainsafe/bls-js/lib/keypair";
 import {MockBeaconApi} from "../../utils/mocks/rpc/beacon";
 import {MockValidatorApi} from "../../utils/mocks/rpc/validator";
 import {ILogger, WinstonLogger} from "../../../src/logger";
+import {IValidatorOptions} from "../../../src/validator/options";
 
 describe('Validator', () => {
   let logger: ILogger = new WinstonLogger();
@@ -27,12 +27,12 @@ describe('Validator', () => {
       validator: new MockValidatorApi(),
     });
 
-    let validatorCtx: ValidatorCtx = {
-      rpc: rpcClient,
+    let validatorCtx: Partial<IValidatorOptions> = {
+      rpcInstance: rpcClient,
       keypair: Keypair.generate(),
     };
 
-    let validator = new Validator(validatorCtx, logger);
+    let validator = new Validator(validatorCtx, {logger});
     await expect(validator.start()).to.not.throw;
     await validator.stop();
   });
